Move exception definitions to source file
