+++ src/exception.cpp
@@ -6,6 +6,8 @@
 
 #include <hpx/hpx_fwd.hpp>
 #include <hpx/config.hpp>
+#include <hpx/error.hpp>
+#include <hpx/error_code.hpp>
 #include <hpx/exception.hpp>
 #include <hpx/state.hpp>
 #include <hpx/version.hpp>
@@ -13,8 +15,11 @@
 #include <hpx/runtime/naming/name.hpp>
 #include <hpx/runtime/threads/threadmanager.hpp>
 #include <hpx/runtime/threads/thread_helpers.hpp>
+#include <hpx/util/assert.hpp>
 #include <hpx/util/backtrace.hpp>
 #include <hpx/util/command_line_handling.hpp>
+#include <hpx/util/filesystem_compatibility.hpp>
+#include <hpx/util/logging.hpp>
 
 #if defined(HPX_WINDOWS)
 #  include <process.h>
@@ -51,6 +56,95 @@ namespace hpx
     bool expect_exception(bool flag)
     {
         return expect_exception_flag.exchange(flag);
+    }
+
+    ///////////////////////////////////////////////////////////////////////////
+    /// Construct a hpx::exception from a \a hpx::error.
+    ///
+    /// \param e    The parameter \p e holds the hpx::error code the new
+    ///             exception should encapsulate.
+    exception::exception(error e)
+      : boost::system::system_error(make_error_code(e, plain))
+    {
+        HPX_ASSERT((e >= success && e < last_error) || (e & system_error_flag));
+        LERR_(error) << "created exception: " << this->what();
+    }
+
+    /// Construct a hpx::exception from a boost#system_error.
+    exception::exception(boost::system::system_error const& e)
+      : boost::system::system_error(e)
+    {
+        LERR_(error) << "created exception: " << this->what();
+    }
+
+    /// Construct a hpx::exception from a \a hpx::error and an error message.
+    ///
+    /// \param e      The parameter \p e holds the hpx::error code the new
+    ///               exception should encapsulate.
+    /// \param msg    The parameter \p msg holds the error message the new
+    ///               exception should encapsulate.
+    /// \param mode   The parameter \p mode specifies whether the returned
+    ///               hpx::error_code belongs to the error category
+    ///               \a hpx_category (if mode is \a plain, this is the
+    ///               default) or to the category \a hpx_category_rethrow
+    ///               (if mode is \a rethrow).
+    exception::exception(error e, char const* msg, throwmode mode)
+      : boost::system::system_error(make_system_error_code(e, mode), msg)
+    {
+        HPX_ASSERT((e >= success && e < last_error) || (e & system_error_flag));
+        LERR_(error) << "created exception: " << this->what();
+    }
+
+    /// Construct a hpx::exception from a \a hpx::error and an error message.
+    ///
+    /// \param e      The parameter \p e holds the hpx::error code the new
+    ///               exception should encapsulate.
+    /// \param msg    The parameter \p msg holds the error message the new
+    ///               exception should encapsulate.
+    /// \param mode   The parameter \p mode specifies whether the returned
+    ///               hpx::error_code belongs to the error category
+    ///               \a hpx_category (if mode is \a plain, this is the
+    ///               default) or to the category \a hpx_category_rethrow
+    ///               (if mode is \a rethrow).
+    exception::exception(error e, std::string const& msg, throwmode mode)
+      : boost::system::system_error(make_system_error_code(e, mode), msg)
+    {
+        HPX_ASSERT((e >= success && e < last_error) || (e & system_error_flag));
+        LERR_(error) << "created exception: " << this->what();
+    }
+
+    /// Destruct a hpx::exception
+    ///
+    /// \throws nothing
+    exception::~exception()
+    {
+    }
+
+    /// The function \a get_error() returns the hpx::error code stored
+    /// in the referenced instance of a hpx::exception. It returns
+    /// the hpx::error code this exception instance was constructed
+    /// from.
+    ///
+    /// \throws nothing
+    error exception::get_error() const HPX_NOEXCEPT
+    {
+        return static_cast<error>(
+            this->boost::system::system_error::code().value());
+    }
+
+    /// The function \a get_error_code() returns a hpx::error_code which
+    /// represents the same error condition as this hpx::exception instance.
+    ///
+    /// \param mode   The parameter \p mode specifies whether the returned
+    ///               hpx::error_code belongs to the error category
+    ///               \a hpx_category (if mode is \a plain, this is the
+    ///               default) or to the category \a hpx_category_rethrow
+    ///               (if mode is \a rethrow).
+    error_code exception::get_error_code(throwmode mode) const HPX_NOEXCEPT
+    {
+        (void)mode;
+        return error_code(this->boost::system::system_error::code().value(),
+            *this);
     }
 }
